+++ magda-scss-compiler/src/renderScss.ts
@@ -1,6 +1,5 @@
 import * as sass from "node-sass";
 import * as cleancss from "clean-css";
-import * as tempy from "tempy";
 import * as fse from "fs-extra";
 import * as escapeStringRegexp from "escape-string-regexp";
 import * as postcss from "postcss";
@@ -99,19 +98,12 @@ export const renderScssFilesExtra = async (
     const varFileContent: string = await fse.readFile(orgVarFile, {
         encoding: "utf-8"
     });
-    const varFile = tempy.file();
-    await fse.writeFile(varFile, replaceParamsFromScss(varFileContent, params));
-    const idxFileContent: string = await fse.readFile(orgIdxfile, {
-        encoding: "utf-8"
-    });
-    const indexFile = tempy.file();
     await fse.writeFile(
-        indexFile,
-        idxFileContent.replace(`@import "variables";`, "")
+        orgVarFile,
+        replaceParamsFromScss(varFileContent, params)
     );
 
-    otherfiles.unshift(indexFile);
-    otherfiles.unshift(varFile);
+    otherfiles.unshift(orgIdxfile);
     return await renderScssData(
         clientRoot,
         otherfiles.map((file: string) => `@import "${file}";`).join("")